Fix Security System.

+++ Systems/SecuritySystem.py
@@ -22,32 +22,36 @@ class SecuritySystem(BaseSystem):
 
 	def _onEnabledChanged(self):
 		BaseSystem._onEnabledChanged(self)
-		self.lastSendTime = datetime.now() - self.sendInterval + self.startDelay
+		self.lastSendTime = datetime.now() + self.startDelay
 
 	def update(self):
 		BaseSystem.update(self)
 		
-		# TODO: add options in UI		
+		# TODO: add options in UI
 		elapsed = datetime.now() - self.lastSendTime
+		print (elapsed)
 		if elapsed > self.sendInterval:
 			if self.activated:
+				InternetService.sendSMS(Config.GSMNumber, "Security Alarm Activated!", "telenor")
 				images = []
 				for i in range(0, self.currImage):
 					images.append("camera" + str(i) + ".jpg")
-				InternetService.sendEMail([Config.EMail], "My Home", "Security Alarm Activated!", images)
-				#InternetService.sendSMS(Config.GSMNumber, "Security Alarm Activated!", "telenor") # TODO: 
-				for i in range(0, self.currImage):
-					os.remove("camera" + str(i) + ".jpg")
+				if InternetService.sendEMail([Config.EMail], "My Home", "Security Alarm Activated!", images): # if send successful
+					for i in range(0, self.currImage):
+						if os.path.isfile("camera" + str(i) + ".jpg"):
+							 os.remove("camera" + str(i) + ".jpg")
+					self.currImage = 0
 			
 			self.activated = False
 		
-		if not self.activated: # if not activated check for motion
+		if not self.activated and elapsed > timedelta(): # if not activated and after delay start - check for motion
 			self.activated = SensorsService.motionDetected()
 			if self.activated:
 				Logger.log("info", "Security System: Activated")
-			self.currImage = 0
 			self.lastSendTime = datetime.now()
+		elif not self.activated:
+			SensorsService.motionDetected()
 		
-		if self.activated and elapsed > self.sendInterval / self.numImages * self.currImage:
+		if self.activated and elapsed > (self.sendInterval / self.numImages) * (self.currImage % (self.numImages + 1)):
 			CameraService.saveImage("camera" + str(self.currImage) + ".jpg", "640x480", 1)
 			self.currImage += 1
